Computer Graphics/Graphics Function Lib/requirements.txt
Pillow==10.1